Fixing bugs
Added stop start to tasks + UI visible
Log frame is displayed with k
Login i done once and there is refresh button now

+++ src/com/age/pinterest/task/PinTask.java
@@ -2,12 +2,13 @@ package com.age.pinterest.task;
 
 import java.io.File;
 import java.util.ArrayList;
+import java.util.concurrent.TimeUnit;
 
 import org.codehaus.jackson.map.ObjectMapper;
 
 import com.age.data.Board;
 import com.age.data.Pin;
-import com.age.data.PinterestAccount;
+import com.age.data.User;
 import com.age.help.BotPaths;
 import com.age.help.FileUtill;
 import com.age.pinterest.api.PinterestApi;
@@ -17,34 +18,43 @@ public class PinTask extends Task {
 	private static final String PINS_LOCATION_URL = BotPaths.ROOT_DIR + "/Users/%s/pins";
 	private final long interval;
 	private Board board;
-	private final PinterestAccount acc;
+	private final User user;
 
-	public PinTask(PinterestAccount acc, Board board, long interval) {
+	public PinTask(User user, Board board, long interval) {
 		this.interval = interval;
 		this.board = board;
-		this.acc = acc;
+		this.user = user;
 
 	}
 
 	@Override
 	public void run() {
-		PinterestApi api = new PinterestApi(acc);
 		ObjectMapper mapper = new ObjectMapper();
-		ArrayList<String> pinFiles = FileUtill.getAllFiles(String.format(PINS_LOCATION_URL, acc.getUser()));
+		PinterestApi api = new PinterestApi(user);
+		ArrayList<String> pinFiles = FileUtill.getAllFiles(String.format(PINS_LOCATION_URL, user.getAccount().getUser()));
 		System.out.println("Pinning to " + board.getName());
 		System.out.println("Interval " + interval);
 		do {
-			try {
-				String filePath = pinFiles.get(0);
-				Pin pin = mapper.readValue(new File(filePath), Pin.class);
-				api.pin(pin, board);
-				new File(filePath).delete();
-				pinFiles.remove(0);
-			} catch (Exception e) {
-				Log.log("Failed to pin  " + e.getMessage());
+			if (this.intervalPassed(interval)) {
+				try {
+					String filePath = pinFiles.get(0);
+					Pin pin = mapper.readValue(new File(filePath), Pin.class);
+					api.pin(pin, board);
+					new File(filePath).delete();
+					pinFiles.remove(0);
+					Log.log("Remaining pins " + pinFiles.size());
+					Log.log("Pin interval is " + TimeUnit.MINUTES.toMinutes(interval) + " minutes");
+				} catch (Exception e) {
+					Log.log("Failed to pin  " + e.getMessage());
+				}
 			}
-		} while (!pinFiles.isEmpty() && this.intervalPassed(interval));
-		Log.log("No more pins for " + acc.getUser());
+		} while (!pinFiles.isEmpty());
+		Log.log("No more pins for " + user.getAccount().getUser());
+	}
+
+	@Override
+	public TaskType getType() {
+		return TaskType.PIN;
 	}
 
 }
